LTK.Decide

Description

Functions used for deciding the complexity class of an automaton. Each complexity class for which these operations are implemented has a separate Decide.class module as well.

Since: 0.2

Basic local classes

isSL :: (Ord e, Ord n) => FSA n e -> Bool

Returns True iff the stringset represented by the given FSA is Strictly Local, that is, if it satisfies Suffix-Substition Closure for some specific factor size \(k\).

isLT :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes an LT stringset.

isLTT :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes an LTT stringset.

Tier-based generalizations

isTSL :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a TSL stringset.

isTLT :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a TLT stringset.

isTLTT :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a TLTT stringset.

Piecewise classes

isSP :: (Ord n, Ord e) => FSA n e -> Bool

Returns True iff the stringset represented by the given FSA is Strictly Piecewise, that is, if the FSA accepts all subsequences of every string it accepts.

isPT :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a PT stringset.

Beyond the Local/Piecewise merge

isSF :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a Star-Free stringset.