LTK.Decide.B

Description

This module implements an algorithm to decide whether a given FSA is in B, the subclass of FO2[<] where all elements are idempotent but the operation is not necessarily commutative. Thus, this is a superclass of CB. The local and tier-local extensions are also provided.

Since: 1.0

Documentation

isB :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a band stringset.

isBM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff the monoid is a band.

isLB :: (Ord n, Ord e) => FSA n e -> Bool

True iff the recognized stringset is locally a band.

isLBM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff the monoid is locally a band.

isTLB :: (Ord n, Ord e) => FSA n e -> Bool

True iff the recognized stringset is locally a band on some tier.

isTLBM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff the monoid is locally a band on some tier.