LTK.Decide.Definite

Description

This module implements an algorithm to decide whether a given FSA is Definite (Def) or Reverse Definite (RDef) based on the classic semigroup characterizations summarized by Brzozowski and Fich in their 1984 work "On Generalized Locally Testable Languages".

Since: 1.0

Plain

isDef :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a definite stringset, characterized by a set of permitted suffixes.

isDefM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff \(Se=e\) for idempotents \(e\).

isRDef :: (Ord n, Ord e) => FSA n e -> Bool

True iff the automaton recognizes a reverse definite stringset, characterized by a set of permitted prefixes.

isRDefM :: (Ord n, Ord e) => SynMon n e -> Bool

True iff \(eS=e\) for idempotents \(e\).

Tier-Based

isTDef :: (Ord n, Ord e) => FSA n e -> Bool

Definite on some tier.

isTDefM :: (Ord n, Ord e) => SynMon n e -> Bool

Definite on the projected subsemigroup.

isTRDef :: (Ord n, Ord e) => FSA n e -> Bool

Reverse definite on some tier.

isTRDefM :: (Ord n, Ord e) => SynMon n e -> Bool

Reverse definite on the projected subsemigroup.