LTK.Decide.SP

Description

This module implements an algorithm to decide whether a given FSA is Strictly Piecewise (SP) by my own direct-construction algorithm.

Since: 0.2

Documentation

isSP :: (Ord n, Ord e) => FSA n e -> Bool

Returns True iff the stringset represented by the given FSA is Strictly Piecewise, that is, if the FSA accepts all subsequences of every string it accepts.