LTK.Factors

Description

This module provides a means to define positive and negative factors over the adjacency or precedence relations, as well as unions and intersections thereof.

Constructions

required :: Factor e -> Literal e

The factor is required to appear in every string. Note that a conjunctive constraint of (required (Substring x True True)) restricts the stringset to at most one word.

forbidden :: Factor e -> Literal e

The factor is not allowed to appear in any word.

buildLiteral :: (Enum n, Ord n, Ord e) => Set e -> Literal e -> FSA n e

Build an FSA representing a single constraint.

build :: (Enum n, NFData n, Ord n, NFData e, Ord e) => Set e -> Set (Conjunction e) -> FSA n e

Build an FSA representing the conjunction of a set of constraints provided in conjunctive normal form.

makeConstraint :: Ord e => [[Literal e]] -> Conjunction e

Combine inner lists by Disjunction, and form a Conjunction of the results.

Logical Expressions

data Factor e

A substring or subsequence, from which to build constraints.

Constructors

Substring

Fields

substring :: [Set e]

The sequence of symbol types, e.g. [wxs0, wxs0] for two consecutive unstressed syllables.

headAnchored :: Bool

Anchored to the head of the word?

tailAnchored :: Bool

Anchored to the tail of the word?

Subsequence [Set e]

data Literal e

A constraint.

Constructors

Literal Bool (Factor e)

data Disjunction e

Multiple constraints, joined by OR.

Constructors

Disjunction (Set (Literal e))

data Conjunction e

Multiple disjunctions, joined by AND.

Constructors

Conjunction (Set (Disjunction e))

Symbols

wnsx is a syllable whose weight is level \(n\) and whose stress is level \(s\). Stress ranges from 0-2, while weight is in theory not limited. Here, only weights up to level 4 are defined. For both weight and stress, "plus" means "greater than zero". For stress, "minus" means "less than two". Arbitrary weight is indicated by using x for \(n\). Arbitrary stress is indicated by omission of sx.

defaultAlphabet :: Set String

Equivalent to wx.

w0s0 :: Set String

Light, unstressed

w0s1 :: Set String

Light, secondary stress

w0s2 :: Set String

Light, primary stress

w0plus :: Set String

Light, some stress

w0minus :: Set String

Light, non-primary stress

w0 :: Set String

Light, any stress

w1s0 :: Set String

Heavy, unstressed

w1s1 :: Set String

Heavy, secondary stress

w1s2 :: Set String

Heavy, primary stress

w1plus :: Set String

Heavy, some stress

w1minus :: Set String

Heavy, non-primary stress

w1 :: Set String

Heavy, any stress

w2s0 :: Set String

Superheavy, unstressed

w2s1 :: Set String

Superheavy, secondary stress

w2s2 :: Set String

Superheavy, primary stress

w2plus :: Set String

Superheavy, some stress

w2minus :: Set String

Superheavy, non-primary stress

w2 :: Set String

Superheavy, any stress

w3s0 :: Set String

Weight 3, unstressed

w3s1 :: Set String

Weight 3, secondary stress

w3s2 :: Set String

Weight 3, primary stress

w3plus :: Set String

Weight 3, some stress

w3minus :: Set String

Weight 3, non-primary stress

w3 :: Set String

Weight 3, any stress

w4s0 :: Set String

Weight 4, unstressed

w4s1 :: Set String

Weight 4, secondary stress

w4s2 :: Set String

Weight 5, primary stress

w4plus :: Set String

Weight 4, some stress

w4minus :: Set String

Weight 4, non-primary stress

w4 :: Set String

Weight 4, any stress

wpluss0 :: Set String

Non-light, unstressed

wpluss1 :: Set String

Non-light, secondary stress

wpluss2 :: Set String

Non-light, primary stress

wplusplus :: Set String

Non-light, some-stress

wplusminus :: Set String

Non-light, non-primary stress

wplus :: Set String

Non-light, any stress

wxs0 :: Set String

Unstressed

wxs1 :: Set String

Secondary stress

wxs2 :: Set String

Primary stress

wxplus :: Set String

Some stress

wxminus :: Set String

Non-primary stress

wx :: Set String

Any weight or stress