LTK.Learn.SP

Description

This module implements a string extension learner for the SP class.

Since: 0.3

Documentation

data SPG a

A representation of an SP grammar.

fSP :: Ord a => Int -> [a] -> SPG a

Return the set of factors under precedence of length \(k\) or less in the given word.