LTK.Porters.Pleb

Description

The (P)iecewise / (L)ocal (E)xpression (B)uilder. This module defines a parser for a representation of logical formulae over subsequence- and adjacency-factors, as well as a mechanism for evaluating (creating an FSA from) the resulting expression tree.

There are two special variables:

it describes the most recent expression, and

universe collects all symbols used.

Documentation

type Dictionary a = Set (String, a)

An association between names and values.

newtype Parse a

The base type for a combinatorial parser.

Constructors

Parse

Fields

doParse :: [Token] -> Either String (a, [Token])

type Env = (Dictionary SymSet, Dictionary Expr, Maybe Expr)

The environment: defined sets of symbols, defined expressions, and possibly a value for the special variable (it).

data Expr

An expression, the root of an expression tree.

type SymSet = Set String

A set of symbols.

data Token

A syntactic unit.

compileEnv :: Env -> Env

Transform all saved expressions into automata to prevent reevaluation.

groundEnv :: Env -> Env

Convert saved automata from descriptions of constraints to descriptions of stringsets. This action effectively removes metadata describing constraint types from the environment.

insertExpr :: Env -> Expr -> Env

Add a new expression to the environment, call it "(it)".

fromAutomaton :: FSA Integer String -> Expr

Generate an expression (sub)tree from an FSA.

fromSemanticAutomaton :: FSA Integer (Maybe String) -> Expr

Generate an expression (sub)tree from an FSA that contains metadata regarding the constraint(s) it represents.

makeAutomaton :: Env -> Maybe (FSA Integer (Maybe String))

Create an FSA from an expression tree and environment, complete with metadata regarding the constraint(s) it represents.

doStatements :: Env -> String -> Env

Parse an input string and update the environment according to the result of the parse.

parseExpr :: Dictionary SymSet -> Dictionary Expr -> Parse Expr

Parse an expression from a Token stream.

readPleb :: String -> Either String (FSA Integer String)

Parse an input string and create a stringset-automaton from the result. If there is an error while parsing, the result is the string "no parse".

restrictUniverse :: Env -> Env

Remove any symbols not present in (universe) from the environment.

tokenize :: String -> [Token]

Convert a string into a stream of tokens ready for parsing.