LTK.Tiers

Description

If an FSA defines a stringset that is the preprojection of some other stringset over a smaller alphabet, the functions in this module can determine what that alphabet is and return the appropriate projective automaton.

Documentation

tier :: (Ord n, Ord e) => FSA n e -> Set e

Determine the tier alphabet for which the given FSA is a preprojection. This could simply be the entire alphabet of the FSA. Precondition: the given FSA must be in normal form.

project :: (Ord n, Ord e) => FSA n e -> FSA n e

Remove symbols not relevant to the given FSA's associated projection (as determined by tier). Precondition: the given FSA must be in normal form.